fastapi/requirements.txt
aiofiles==0.6.0
bcrypt==3.2.0
boto3==1.17.19
botocore==1.20.19
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
jmespath==0.10.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
